Move middlewares for consistent redux setup
